Add files via upload
This arm was designed to fit the (LD-260MG) type of engine with the appropriate dimensions. Also, cooling holes were placed, one upper to expel the hot air and the lower one to enter the cold air (heat transfer course) and a hole was added for engine wire in addition to the presence of a cover design that was also attached.